+++ R/helpers_ggbetween_t_subtitles.R
@@ -654,11 +654,15 @@ subtitle_t_bayes <- function(data,
 #'
 #' References-
 #' \itemize{
-#'   \item Cooper, Harris, Hedges, Larry V., Valentine, Jeffrey C., *The Handbook of Research Synthesis and Meta-Analysis*, 2009.
-#'   \item Cumming, G., Finch, S., *A Primer On The Understanding, Use, And Calculation Of Confidence Intervals That Are Based On Central And Noncentral Distributions*, Educational and Psychological Measurement, Vol. 61 No. 4,
-#' August 2001 532-574.
-#'   \item Cohen, J. (1988). *Statistical power analysis for the behavioral sciences (2nd ed.)* Hillsdale, NJ: Lawrence Erlbaum Associates.
-#'   \item David C. Howell (2010). *Confidence Intervals on Effect Size* retrieved from (\url{https://www.uvm.edu/~dhowell/methods7/Supplements/Confidence\%20Intervals\%20on\%20Effect\%20Size.pdf}).
+#' \item Cooper, Harris, Hedges, Larry V., Valentine, Jeffrey C., The Handbook
+#' of Research Synthesis and Meta-Analysis, 2009. \item Cumming, G., Finch, S.,
+#' A Primer On The Understanding, Use, And Calculation Of Confidence Intervals
+#' That Are Based On Central And Noncentral Distributions, Educational and
+#' Psychological Measurement, Vol. 61 No. 4, August 2001 532-574. \item Cohen,
+#' J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.)
+#' Hillsdale, NJ: Lawrence Erlbaum Associates. \item David C. Howell (2010).
+#' Confidence Intervals on Effect Size, retrieved from
+#' (\url{https://www.uvm.edu/~dhowell/methods7/Supplements/Confidence\%20Intervals\%20on\%20Effect\%20Size.pdf}).
 #' }
 #'
 #' @examples
@@ -732,7 +736,7 @@ effsize_t_parametric <- function(formula = NULL,
     df <- length(x) - 1
     mean.diff <- mean(x) - mu
     d <- mean.diff / sd.est
-    Sigmad <- sqrt((n /(n / 2) ^ 2 ) + (d ^ 2 / (2 * n)))
+    Sigmad <- sqrt((n / (n / 2)^2) + (d^2 / (2 * n)))
     Z <- -stats::qt((1 - conf.level) / 2, df)
     lower.ci <- c(d - Z * Sigmad)
     upper.ci <- c(d + Z * Sigmad)